create FormValidator service and remove unused AbstractFactory
