+++ src/UI/Factory/UpdateEntityFactory.php
@@ -2,6 +2,7 @@
 
 namespace App\UI\Factory;
 
+use App\Application\Validation\FormValidator;
 use App\Domain\Entity\Interfaces\EntityInterface;
 use App\Domain\Model\Interfaces\ModelInterface;
 use App\Domain\Repository\Interfaces\Manageable;
@@ -11,7 +12,6 @@ use Doctrine\Common\Persistence\ObjectRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 /**
  * Class UpdateEntityFactory
@@ -33,16 +33,24 @@ final class UpdateEntityFactory
     private $formFactory;
 
     /**
+     * @var FormValidator
+     */
+    private $formValidator;
+
+    /**
      * UpdateEntityFactory constructor.
      * @param EntityManagerInterface $entityManager
      * @param FormFactoryInterface $formFactory
+     * @param FormValidator $formValidator
      */
     public function __construct(
         EntityManagerInterface $entityManager,
-        FormFactoryInterface $formFactory
+        FormFactoryInterface $formFactory,
+        FormValidator $formValidator
     ) {
         $this->entityManager = $entityManager;
         $this->formFactory = $formFactory;
+        $this->formValidator = $formValidator;
     }
 
     /**
@@ -60,6 +68,10 @@ final class UpdateEntityFactory
 
         $form = $this->formFactory->create($model->getEntityType(), $entity);
         $this->processForm($request, $form);
+        if ($form->isSubmitted() && !$form->isValid()) {
+            $this->formValidator->throwValidationErrorException($form);
+        }
+
         $repository->saveOrUpdate($entity, true);
 
         return $model::createFromEntity($entity);
